requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aiounittest==1.3.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.30.1
azure-cosmos==3.2.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.9.1
backcall==0.2.0
blinker==1.8.1
botbuilder==0.0.1
botbuilder-azure==4.15.0
botbuilder-core==4.15.0
botbuilder-dialogs==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-sangam-twilio-whatsapp-adapter==0.4.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
cryptography==36.0.1
dataclasses-json==0.5.9
datedelta==1.4
debugpy==1.6.7
decorator==5.1.1
emoji==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
faiss-cpu==1.7.4
filetype==1.2.0
firebase-admin==6.5.0
Flask==3.0.3
frozenlist==1.3.3
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grapheme==0.6.0
greenlet==2.0.2
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.3
joblib==1.3.1
jsonpickle==1.4.2
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.137
langchainplus-sdk==0.0.20
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
msal==1.22.0
msg-parser==1.2.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.4
multipledispatch==1.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
oauthlib==3.2.2
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
priority==2.0.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyngrok==6.0.0
pypandoc==1.11
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
Quart==0.19.5
recognizers-date-time==1.0.0a1
recognizers-number==1.0.0a3
recognizers-number-with-unit==1.0.0a1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.49
stack-data==0.6.2
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twilio==7.3.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.9.0
unstructured==0.8.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.29.0
waitress==3.0.0
wcwidth==0.2.6
Werkzeug==3.0.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.4
zipp==3.16.2
